Before gutting printing
PPD logic has been purged, but printing needs to be completely changed. We need Text, Pdf, Raw, and possibly PostScript engines that take a chunk of data and create a temporary file of the appropriate type then the temp needs to be stuck into a print job. No more paginating one at a time to a printer.